--- src/main/java/hometoogether/hometoogether/domain/trial/repository/TrialRepository.java
@@ -1,7 +1,0 @@
-package hometoogether.hometoogether.domain.trial.repository;
-
-import hometoogether.hometoogether.domain.trial.domain.Trial;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TrialRepository extends JpaRepository<Trial, Long> {
-}
